requirements.txt
feedparser==6.0.11
pandas==2.2.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rapidfuzz==3.9.6
requests==2.32.3